SAA in progress

+++ src/org/srg/scpp_im/strategy/OSSCPP_TargetMV.java
@@ -122,7 +122,7 @@ public class OSSCPP_TargetMV extends SelfConfirmingPricePrediction {
 			}
 			for (BitSet bs : bitVector)
 			{
-				int value = typeDist.get(bs).intValue();
+				double value = (double)typeDist.get(bs).intValue();
 				double cost = 0.0;
 				for (int j=0;j<bs.length();j++)
 				{
